Fixed parse error in scenario
ignored return from 20-sim launch (windows xp problem)

git-svn-id: svn://svn.code.sf.net/p/destecs/code/trunk@150 fd54fc82-ac30-4e97-bb06-279643433f24

+++ ide/core/src/main/java/org/destecs/ide/core/internal/builder/IncrementalProjectBuilder.java
@@ -221,10 +221,12 @@ public class IncrementalProjectBuilder extends
 	@SuppressWarnings("unchecked")
 	protected Object parse(ParserWrapper parser, IFile file) throws Exception
 	{
-		FileUtility.deleteMarker(file, IMarker.PROBLEM, IDestecsCoreConstants.PLUGIN_ID);
-
 		Object result = parser.parse(file.getLocation().toFile());
+		
+		if (file == null || !file.exists()|| !file.isAccessible())
+			return result;
 
+		FileUtility.deleteMarker(file, IMarker.PROBLEM, IDestecsCoreConstants.PLUGIN_ID);
 		if (parser.hasErrors())
 		{
 			for (Object err : parser.getErrors())
